feat(review): remove security boundary change signals implementation
